Version 1.7.9 1.21.11, Fixed inverse third person bug, fixed all warnings, removed unused variables and files, fixed deprecated function calls.

+++ src/main/java/net/hiddenhally/buganair/mixin/EntityGliderRotationMixin.java
@@ -51,6 +51,6 @@ public abstract class EntityGliderRotationMixin {
         float i = MathHelper.sin(g);
         float j = MathHelper.cos(f);
         float k = MathHelper.sin(f);
-        return new Vec3d((double)(i * j), (double)(-k), (double)(h * j));
+        return new Vec3d(i * j, -k, h * j);
     }
 }
